0x12 and 0x42 tests and refactor starthttps test helper to separate file
ChangeLog:none
